chore: update code to follow go packaging conventions
