+++ media.go
@@ -3,7 +3,18 @@ package goaccount
 import (
 	"encoding/json"
 	"mime/multipart"
+	"time"
+
+	"github.com/google/uuid"
 )
+
+type Media struct {
+	ID         uuid.UUID `db:"id" json:"id"`
+	IdentityID uuid.UUID `db:"identity_id" json:"identity_id"`
+	URL        string    `db:"url" json:"url"`
+	Filename   string    `db:"filename" json:"filename"`
+	CreatedAt  time.Time `db:"created_at" json:"created_at"`
+}
 
 func UploadMedia(file multipart.File, media interface{}) error {
 	response, err := RequestMultipart(RequestOptions{
