+++ organization.go
@@ -2,7 +2,45 @@ package goaccount
 
 import (
 	"encoding/json"
+	"time"
+
+	"github.com/google/uuid"
+	"github.com/jmoiron/sqlx/types"
 )
+
+type Organization struct {
+	ID          uuid.UUID `db:"id" json:"id"`
+	Shortname   string    `db:"shortname" json:"shortname"`
+	Name        *string   `db:"name" json:"name"`
+	Bio         *string   `db:"bio" json:"bio"`
+	Description *string   `db:"description" json:"description"`
+	Email       *string   `db:"email" json:"email"`
+	Phone       *string   `db:"phone" json:"phone"`
+
+	City    *string `db:"city" json:"city"`
+	Country *string `db:"country" json:"country"`
+	Address *string `db:"address" json:"address"`
+	Website *string `db:"website" json:"website"`
+
+	Mission *string `db:"mission" json:"mission"`
+	Culture *string `db:"culture" json:"culture"`
+
+	LogoID   *uuid.UUID     `db:"logo_id" json:"logo_id"`
+	Logo     *Media         `db:"-" json:"logo"`
+	LogoJson types.JSONText `db:"logo" json:"-"`
+
+	CoverID   *uuid.UUID     `db:"cover_id" json:"cover_id"`
+	Cover     *Media         `db:"-" json:"cover"`
+	CoverJson types.JSONText `db:"cover" json:"-"`
+
+	Status OrganizationStatusType `db:"status" json:"status"`
+
+	VerifiedImpact bool `db:"verified_impact" json:"verified_impact"`
+	Verified       bool `db:"verified" json:"verified"`
+
+	CreatedAt time.Time `db:"created_at" json:"created_at"`
+	UpdatedAt time.Time `db:"updated_at" json:"updated_at"`
+}
 
 func GetAllOrganizations() ([]Organization, error) {
 	response, err := Request(RequestOptions{
